Updates to the hdl32e gps imu node time sync
This commit attempts to update the hdl32e GPS/IMU node started by
ddillenberger.  It updates the time offset calculations between the
velodyne's time-since-hour counter and the GPS NMEA string to provide
the correct reconstructed timestamp.

It also changes most of the message types.  The sensor scan time is now
published as a TimeReference message, so that it can be correlated with
the imu and temperature messages via the publish time.  It now publishes
the NMEA string as a nmea_msgs/Sentence message for easier processing by
the navsat drivers.  It also changes the imu publishing from the
geometry_msgs/TwistStamped message to sensor_msgs/Imu.