Updating .travis.yml after testing. Lowering 64e Hz test threshold.
This build is almost a full minute faster across both environments,
uses the "standardized" industrial_ci script, and has been updated
to build in Indigo and Kinetic instead of Indigo and Jade.

However, due to the docker-in-docker performance hit, the 64e test
failed to acheive the 5Hz publish rate. Lowering the error threshold
to 3.0Hz fixed this but I'm not sure how everyon feels about
this change.