driver: add enabled reconfigure param
This will allow polling to be disabled when the device has been disabled
in some way. Specifically this is interesting when the rpm is set to 0
when not is use. This also adds a mask value for the time_offset so that
it won't be touched if it wasn't reconfigured.