Fix the target frame ID.
This essentially fixes the problem that is currently on master
and reported in https://github.com/ros-drivers/velodyne/issues/249.
Here, we take an empty target frame (the default) to mean that
we should use the frame_id incoming from the VelodyneScan;
any other value (which can be set via a parameter), and we
override the frame_id with that.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>