Fix dead pointers in calibration.cc.
Both min_intensity and max_intensity create a temporary object,
and then store a pointer to the object.  The object then goes
out of scope, and the pointer is dereferenced.  Fix this by
just not using the temporary at all and instead doing the
float conversion directly, which is less code and more correct.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>